Reintroduced usage of increments in migrations and updated composer lock
*Reintroduced the usage of increments('id) to maintain unsigned integer
- prevents issues when using integrated with schooldata application tables
*Updated composer.json and composer.lock

+++ src/LaraQuiz/database/migrations/2017_07_04_123900_create_quiz_runner_survey_tag_table.php
@@ -1,8 +1,6 @@
 <?php
 declare(strict_types=1);
 
-
-
 use Illuminate\Database\Migrations\Migration;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Support\Facades\Schema;
@@ -35,8 +33,8 @@ class CreateQuizRunnerSurveyTagTable extends Migration
         Schema::create($this->table, function (Blueprint $table) {
             $table->engine = 'InnoDB';
 
-            $table->unsignedBigInteger('survey_id');
-            $table->unsignedBigInteger('tag_id');
+            $table->unsignedInteger('survey_id');
+            $table->unsignedInteger('tag_id');
 
             $table->primary(['survey_id', 'tag_id']);
